Move atmosphere uniforms out of AtmosphericScatteringPostProcess
